+++ src/engine/sprite_rendering_system.hpp
